Enum gen_epub_book::Error

pub enum Error {
    Io {
        desc: &'static str,
        op: &'static str,
        more: Option<&'static str>,
    },
    Parse {
        tp: &'static str,
        wher: &'static str,
        more: Option<&'static str>,
    },
    FileNotFound {
        who: &'static str,
        path: PathBuf,
    },
    WrongFileState {
        what: &'static str,
        path: PathBuf,
    },
    WrongElementAmount {
        element: &'static str,
        actual: usize,
        relation: &'static str,
        bound: usize,
    },
    RequiredElementMissing(&'static str),
}

Enum representing all possible ways the application can fail.

Variants
Io

An I/O error occured.

This includes higher-level I/O errors like FS ones.

Fields of Io

desc: &'static str

The file the I/O operation regards.

op: &'static str

The failed operation.

This should be lowercase and imperative ("create", "open").

more: Option<&'static str>

Additional data.

Parse

A parsing error occured.

Fields of Parse

tp: &'static str

What failed to parse.

Something like "URL", "datetime".

wher: &'static str

Where the thing that failed to parse would go, were it to parese properly.

more: Option<&'static str>

Additional data.

FileNotFound

A requested file doesn't exist.

Fields of FileNotFound

who: &'static str

What requested the file.

path: PathBuf

The file that should exist.

WrongFileState

A path is in a wrong state.

Fields of WrongFileState

what: &'static str

What the file is not.

path: PathBuf

The file that should be.

WrongElementAmount

An incorrect amount of book elements.

Fields of WrongElementAmount

element: &'static str

The element's name.

actual: usize

Current amount.

relation: &'static str

How it should be.

bound: usize

What it should be.

RequiredElementMissing(&'static str)

A required book element is missing.

Methods

impl Error

pub fn print_error<W: Write>(&self, err_out: &mut W)

Write the error message to the specified output stream.

Examples

let mut out = Vec::new();
Error::Io {
    desc: "network",
    op: "write",
    more: Some("full buffer"),
}.print_error(&mut out);
assert_eq!(String::from_iter(out.iter().map(|&i| i as char)),
           "Writing network failed: full buffer.\n".to_string());

pub fn exit_value(&self) -> i32

Get the executable exit value from an Error instance.

Examples

assert_eq!(Error::Io {
    desc: "",
    op: "",
    more: None,
}.exit_value(), 1);