Module gen_epub_book::ops

Main functions doing actual work.

First, use parse_descriptor() to get an element list from a descriptor file, then construct an EPubBook::from_elements(), validate and absolutise paths via EPubBook::normalise_paths() and write the book with EPubBook::write_zip().

Structs

EPubBook
Full ePub book, bundled together.

Enums

BookElement
A single element of the e-book

EPubContentType
Enum representing what can go in an ePub

Functions

find_title
Find an ePub title line in the specified input stream.

parse_descriptor
Parse the whole descriptor, stopping at the first encountered error

Type Definitions

EPubData
An (ID, filename, data) tuple that represents a single data point of an ePub.